+++ packages/jest-diff/src/normalizeDiffOptions.ts
@@ -6,7 +6,7 @@
  */
 
 import chalk = require('chalk');
-import type {DiffOptions, DiffOptionsNormalized} from './types';
+import type {CompareKeys, DiffOptions, DiffOptionsNormalized} from './types';
 
 export const noColor = (string: string): string => string;
 
@@ -24,6 +24,7 @@ const OPTIONS_DEFAULT: DiffOptionsNormalized = {
   commonColor: chalk.dim,
   commonIndicator: ' ',
   commonLineTrailingSpaceColor: noColor,
+  compareKeys: undefined,
   contextLines: DIFF_CONTEXT_DEFAULT,
   emptyFirstOrLastLinePlaceholder: '',
   expand: true,
@@ -31,6 +32,11 @@ const OPTIONS_DEFAULT: DiffOptionsNormalized = {
   omitAnnotationLines: false,
   patchColor: chalk.yellow,
 };
+
+const getCompareKeys = (compareKeys?: CompareKeys): CompareKeys =>
+  compareKeys && typeof compareKeys === 'function'
+    ? compareKeys
+    : OPTIONS_DEFAULT.compareKeys;
 
 const getContextLines = (contextLines?: number): number =>
   typeof contextLines === 'number' &&
@@ -45,5 +51,6 @@ export const normalizeDiffOptions = (
 ): DiffOptionsNormalized => ({
   ...OPTIONS_DEFAULT,
   ...options,
+  compareKeys: getCompareKeys(options.compareKeys),
   contextLines: getContextLines(options.contextLines),
 });